backend-python-fastapi/requirements.txt
uvicorn==0.21.1
fastapi==0.95.1
python-jose==3.3.0
passlib==1.7.4
sqlalchemy==2.0.12
alembic==1.11.0
#psycopg2==2.9.6
psycopg2-binary==2.9.10